Cody: Alternative implementation for markdown escaping
